Fixed issue with ios ellipses not being drawn on mutable images.  https://github.com/codenameone/CodenameOne/issues/1878

+++ Ports/iOSPort/nativeSources/CodenameOne_GLViewController.m
@@ -741,7 +741,7 @@ CGContextRef drawArc(CGContextRef context, int color, int alpha, int x, int y, i
         CGMutablePathRef path = CGPathCreateMutable();
         
         CGAffineTransform t = CGAffineTransformMakeTranslation(cx, cy);
-        t = CGAffineTransformConcat(CGAffineTransformMakeScale(1.0, height/width), t);
+        t = CGAffineTransformConcat(CGAffineTransformMakeScale(1.0, height/(float)width), t);
         
         CGFloat radius = width/2;
         if (fill){
